+++ build.gradle
@@ -14,6 +14,10 @@ repositories {
 }
 
 dependencies {
+    implementation group: 'ca.umontreal.iro.simul', name: 'ssj', version: '3.3.2'
+    testImplementation "junit:junit:4.13.2"
+    testRuntimeOnly "org.junit.vintage:junit-vintage-engine:5.7.0"
+    compileOnly group: 'org.projectlombok', name: 'lombok', version: '1.18.28'
 }
 
 tasks.named('test') {
