+++ src/main/java/melsongit/com/github/wrestle/Match.java
@@ -1,0 +1,36 @@
+package melsongit.com.github.wrestle;
+
+public class Match {
+}
+
+//class Match
+//{
+//    public:
+//    Match(Wrestler& wr1, Wrestler& wr2) : w1(wr1), w2(wr2)
+//    {
+//        float sigma = max(abs(w1.abilityScore - w2.abilityScore) / 3.f, 15.f);
+//
+//        std::random_device rd;
+//
+//        std::mt19937 e2(rd());
+//
+//        std::normal_distribution<double> distribution1(w1.abilityScore, sigma);
+//        std::normal_distribution<double> distribution2(w2.abilityScore, sigma);
+//        float score1 = distribution1(e2);
+//        float score2 = distribution2(e2);
+//        if (score1 > score2)
+//        {
+//            winner = w1;
+//            looser = w2;
+//        }
+//        else
+//        {
+//            winner = w2;
+//            looser = w1;
+//        }
+//    }
+//    Wrestler w1;
+//    Wrestler w2;
+//    Wrestler winner;
+//    Wrestler looser;
+//};
