+++ src/main/java/melsongit/com/github/wrestle/utils/FightRecord.java
@@ -1,0 +1,12 @@
+package melsongit.com.github.wrestle.utils;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class FightRecord {
+
+    int wins = 0;
+    int losses = 0;
+}
