+++ src/main/java/melsongit/com/github/wrestle/utils/WeightClasses.java
@@ -1,0 +1,20 @@
+package melsongit.com.github.wrestle.utils;
+
+public class WeightClasses {
+
+    static final int[] weightClassCategory = {
+            94,
+            106,
+            113,
+            120,
+            126,
+            132,
+            138,
+            145,
+            152,
+            167,
+            189,
+            220,
+            285
+    };
+}
