+++ src/test/java/melsongit/com/github/wrestle/MainApplicationTest.java
@@ -1,0 +1,10 @@
+package melsongit.com.github.wrestle;
+
+import org.junit.Test;
+
+public class MainApplicationTest {
+
+    @Test
+    public void contextLoads() {
+    }
+}
